Add workspace for general utility functionality
Created the util/ directory with a new Cargo.toml to describe it.
Note that no dependencies are currently required, so the manifest is
sparse.
Added a (deliberately) very limited set of math features, and
set up workspace for re-exporting through lib.rs.
Updated outer Cargo.toml to include the new workspace.